+++ docs/conf.py
@@ -98,7 +98,8 @@ pygments_style = 'sphinx'
 # -- Options for HTML output ---------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
+# a list of builtin themes. Currently Setting it to default to allow RTD 
+# builds to pickup the new default theme.
 html_theme = 'default'
 
 # Theme options are theme-specific and customize the look and feel of a theme